Visual fixes and added details
Removed the sidebar and added it to the pop-out menu. I don't think it
was adding anything to the site where it was, and I'd like more emphasis
on it.

Added portions of the file system to ignore since my troubleshooting
folder isn't helpful to others so much.

Also visual tweaks.